feat(web/auth): extract sign-out into an overridable seam
AuthService.signOut hard-coded a Django session-logout form POST, so a
downstream edition with a different logout (e.g. a hosted OIDC edition
whose sign-out is a JSON request returning an upstream logout URL) had to
fork the whole service to change one method.

Move the form-POST logic into core/services/auth-signout.ts behind a
stable performSignOut(client, baseUrl) export and have signOut delegate
to it — mirroring the existing core/hooks/oauth/extended.tsx seam.
Self-hosted behavior is unchanged; editions now override just this file.
